chore(build): add missing export
