+++ src/muiComponents/Text/index.ts
@@ -1,1 +1,2 @@
-export { default, TextProps } from './Text';
+export { default } from './Text';
+export { TextProps } from './TextConfig';
