+++ src/pages/Version/index.ts
@@ -1,8 +1,4 @@
-export {
-  DetailContext,
-  DetailContextConsumer,
-  DetailContextProvider,
-  DetailContextProps,
-  VersionPageConsumerProps,
-} from './context';
+export { DetailContext, DetailContextConsumer, DetailContextProvider } from './context';
+export { VersionPageConsumerProps, DetailContextProps } from './version-config';
+
 export { default } from './Version';
